ProdlerBackend/requirements.txt
asgiref==3.4.1
Django==3.2.7
django-redis==5.1.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
psycopg2==2.9.1
PyJWT==2.1.0
pytz==2021.1
redis==3.5.3
sqlparse==0.4.2
